backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-telegram-bot==21.4
python-dotenv==1.0.0
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0.35
asyncpg
alembic
python-dotenv
PyJWT
cryptography
python-telegram-bot>=21.0